+++ core/src/main/java/com/novoda/github/reports/data/model/Repository.java
@@ -5,11 +5,11 @@ import com.google.auto.value.AutoValue;
 @AutoValue
 public abstract class Repository {
 
-    public static Repository create(Integer id, String name, boolean isPrivate) {
+    public static Repository create(Long id, String name, boolean isPrivate) {
         return new AutoValue_Repository(id, name, isPrivate);
     }
 
-    public abstract Integer id();
+    public abstract Long id();
 
     public abstract String name();
